+++ src/main/java/org/compiere/orm/X_AD_User_OrgAccess.java
@@ -51,7 +51,7 @@ public class X_AD_User_OrgAccess extends PO implements UserOrganizationalAccess 
      * @return User within the system - Internal or Business Partner Contact
      */
     public int getUserId() {
-        Integer ii = (Integer) getValue(COLUMNNAME_AD_User_ID);
+        Integer ii = getValue(COLUMNNAME_AD_User_ID);
         if (ii == null) return 0;
         return ii;
     }
